[#221] rename: DTO 이름 변경 및 import문 최신화
